+++ exts/base.py
@@ -18,6 +18,7 @@ import discord.ext.commands as commands
 
 import info
 import common as cmn
+from data import options as opt
 
 
 class QrmHelpCommand(commands.HelpCommand):
@@ -120,6 +121,10 @@ class BaseCog(commands.Cog):
                         self.commit = bf.readline().strip()[:7]
         else:
             self.commit = ""
+        self.donation_links = {
+            "Ko-Fi": "https://ko-fi.com/miaowware",
+            "LiberaPay": "https://liberapay.com/miaowware",
+        }
 
     @commands.command(name="info", aliases=["about"])
     async def _info(self, ctx: commands.Context):
@@ -194,6 +199,17 @@ class BaseCog(commands.Cog):
                             All issues and requests related to resources (including maps, band charts, data) \
                             should be added in \
                             [miaowware/qrm-resources](https://github.com/miaowware/qrm-resources/issues)."""
+        await ctx.send(embed=embed)
+
+    @commands.command(name="donate")
+    async def _donate(self, ctx: commands.Context):
+        """Shows ways to help support development of the bot via donations."""
+        embed = cmn.embed_factory(ctx)
+        embed.title = "Help Support qrm's Development!"
+        embed.description = ("Donations are always appreciated, and help with server and infrastructure costs."
+                             "\nThank you for your support!")
+        for title, url in self.donation_links.items():
+            embed.add_field(name=title, value=url, inline=False)
         await ctx.send(embed=embed)
 
     @commands.command(name="echo", aliases=["e"], category=cmn.cat.admin)