+++ templates/data/options.py
@@ -46,6 +46,7 @@ exts = [
     "propagation",
 ]
 
+
 # Either "time", "random", or "fixed" (first item in statuses)
 status_mode = "fixed"